+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://gitlab.heigit.org/climate-action/api-gateway/-/compare/2.6.0...main)
+## [Unreleased](https://gitlab.heigit.org/climate-action/api-gateway/-/compare/2.7.0...main)
+
+## [2.7.0](https://gitlab.heigit.org/climate-action/api-gateway/-/releases/2.7.0) - 2026-06-26
+
+### Added
+
+- update the underlying `climatoology` library to 7.2.0 enabeling planned plugins (and a display file bugfix)
 
 ## [2.6.0](https://gitlab.heigit.org/climate-action/api-gateway/-/releases/2.6.0) - 2026-06-03